requirements.txt
streamlit>=1.20
tensorflow>=2.12
tensorflow-hub>=0.12
Pillow
numpy
